+++ modular-synth-backend/src/sc-controller/SuperColliderPacketBuilder.cpp
@@ -1,6 +1,6 @@
 #include "sc-controller/SuperColliderPacketBuilder.hpp"
 
-void SuperColliderPacketBuilder::s_new( std::string definition, int synth, int action, int target, std::vector<std::pair<std::variant<int, std::string>, std::variant<float, int, std::string>>> control)
+void SuperColliderPacketBuilder::s_new( std::string definition, int synth, int action, int target, std::vector<std::pair<std::variant<int, std::string>, std::variant<int, float, std::string>>> control)
 {
     *this << osc::BeginMessage("/s_new");
     *this << definition.c_str();
